BUILD: Moved header files to subdirectory 'include' and sources files to subdirectory 'src'. This allows us to identify the include directory more clearly in the build process. Also grouped library files into a subdirectory that has the same name as the library. Use <>-notation i.o. ""-notation for includes, so that we only look in the compiler-specified include directories.
