--- fltk.h
@@ -1,17 +1,0 @@
-#ifndef FLTK_GUARD
-#define FLTK_GUARD 1
-
-#include "FL/Fl.H"
-#include "FL/Fl_Window.H" 
-#include "FL/Fl_Button.H" 
-#include "FL/Fl_Input.H" 
-#include "FL/Fl_Output.H" 
-#include <cstdlib>       // for exit(0)
-#include "FL/fl_draw.H"
-#include "FL/Enumerations.H"
-
-#include "Fl/Fl_JPEG_Image.H"
-#include "Fl/Fl_GIF_Image.H"
-
-#endif
-
