+++ include/fltk.hpp
@@ -1,0 +1,15 @@
+#ifndef FLTK_GUARD
+#define FLTK_GUARD 1
+
+#include <cstdlib>       // for exit(0)
+#include <FL/Enumerations.H>
+#include <FL/Fl.H>
+#include <FL/Fl_Button.H> 
+#include <FL/fl_draw.H>
+#include <Fl/Fl_GIF_Image.H>
+#include <FL/Fl_Input.H> 
+#include <Fl/Fl_JPEG_Image.H>
+#include <FL/Fl_Output.H> 
+#include <FL/Fl_Window.H> 
+
+#endif
